대전 게임 서버 수정

+++ OmokGameServer/OmokBinaryRequestInfo.cs
@@ -4,10 +4,10 @@ namespace OmokGameServer
 {
     public class OmokBinaryRequestInfo : BinaryRequestInfo
     {
-        public Int16 Size { get; private set; }
-        public Int16 PacketId { get; private set; }
+        public int Size { get; private set; }
+        public int PacketId { get; private set; }
 
-        public OmokBinaryRequestInfo(Int16 size, Int16 packetId, byte[] body)
+        public OmokBinaryRequestInfo(int size, int packetId, byte[] body)
             :base(null, body)
         {
             this.Size = size;
