+++ OmokGameServer/ServerPacketData.cs
@@ -2,12 +2,12 @@ namespace OmokGameServer
 {
     public class ServerPacketData
     {
-        public Int16 packetSize;
+        public int packetSize;
         public string sessionId;
-        public Int16 packetId;
+        public int packetId;
         public byte[] bodyData;
 
-        public void SetPacket(string sessionId, Int16 packetId, byte[] bodyData)
+        public void SetPacket(string sessionId, int packetId, byte[] bodyData)
         {
             this.sessionId = sessionId;
             this.packetId = packetId;
@@ -18,7 +18,7 @@ namespace OmokGameServer
             }
         }
 
-        public void SetPacketNoBody(string sessionId, Int16 packetId)
+        public void SetPacketNoBody(string sessionId, int packetId)
         {
             this.sessionId = sessionId;
             this.packetId = packetId;
